use correct types in std::min

+++ libosmscout-import/src/osmscout/import/WaterIndexProcessor.cpp
@@ -1240,8 +1240,8 @@ namespace osmscout {
 
     progress.Info("Filter intersecting islands");
 
-    for (size_t i=0; i<std::min(transformedCoastlines.size(), 100lu); i++) {
-      progress.SetProgress(i,std::min(transformedCoastlines.size(), 100lu));
+    for (size_t i=0; i<std::min(transformedCoastlines.size(), static_cast<size_t>(100)); i++) {
+      progress.SetProgress(i,std::min(transformedCoastlines.size(), static_cast<size_t>(100)));
 
       for (size_t j=i+1; j<transformedCoastlines.size(); j++) {
         assert(i!=j);
